+++ media/welcome.html
@@ -1,6 +1,7 @@
 <html>
 
 <head>
+    <meta charset="utf-8">
     <link href="{0}" rel="stylesheet">
     <script src="{1}/bootstrap.bundle.min.js"></script>
     <script src="{1}/fontawesome.js"></script>
