+++ src/extension.ts
@@ -116,6 +116,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	if (Container.universal.hasConnection()) {
 		if (await Container.universal.waitForAlive()) {
+			await Container.universal.connectDebugger();
 			await Container.universal.installAndLoadModule();
 		}
 	}
