+++ walkthroughs/getting-started/2-login.md
@@ -1,5 +1,5 @@
 ## Login to PowerShell Universal
 
-Once you have installed PowerShell Universal, it will be listening at `http://localhost:5000`. You can login with the username `admin` and the password `admin`. 
+Once you have installed PowerShell Universal, it will be listening at `http://localhost:5000`. You will have to set your default user name and password and then can login.
 
 [💡 Login to PowerShell Universal](http://localhost:5000)
